[共通] LogLevel, LogType
